feat: Extract domain modules and types (Phase 2)
- Created domain-specific modules in src/domains/:
  - base.ts: Base class for all domain modules
  - projects.ts: Project operations
  - issues.ts: Issue operations with complex filtering
  - metrics.ts: Metrics operations
  - measures.ts: Measures operations with pagination
  - system.ts: System health/status operations
  - quality-gates.ts: Quality gate operations
  - source-code.ts: Source code operations
  - hotspots.ts: Security hotspot operations

- Extracted all type definitions to src/types/:
  - common.ts: Shared types (SeverityLevel, PaginationParams)
  - Separate files for each domain's types
  - index.ts: Central re-export maintaining backward compatibility

- Refactored SonarQubeClient to use facade pattern:
  - Delegates all operations to domain modules
  - Maintains backward compatibility
  - Supports all authentication methods (token, basic, passcode)

- Fixed API compatibility issues:
  - Updated builder method calls to match sonarqube-web-api-client API
  - Fixed pagination handling in measures domain
  - Handled deprecated hotspots parameter in issues

All tests passing, TypeScript checks pass, and code is properly formatted.

Part of #104: Refactor codebase into smaller, focused modules
