+++ src/domains/index.ts
@@ -1,0 +1,10 @@
+// Export all domain modules
+export { BaseDomain } from './base.js';
+export { ProjectsDomain } from './projects.js';
+export { IssuesDomain } from './issues.js';
+export { MetricsDomain } from './metrics.js';
+export { MeasuresDomain } from './measures.js';
+export { SystemDomain } from './system.js';
+export { QualityGatesDomain } from './quality-gates.js';
+export { SourceCodeDomain } from './source-code.js';
+export { HotspotsDomain } from './hotspots.js';
