+++ src/sonarqube.ts
@@ -1,12 +1,100 @@
-import { SonarQubeClient as WebApiClient, MeasuresAdditionalField } from 'sonarqube-web-api-client';
+import { SonarQubeClient as WebApiClient } from 'sonarqube-web-api-client';
 import { createLogger } from './utils/logger.js';
+import {
+  ProjectsDomain,
+  IssuesDomain,
+  MetricsDomain,
+  MeasuresDomain,
+  SystemDomain,
+  QualityGatesDomain,
+  SourceCodeDomain,
+  HotspotsDomain,
+} from './domains/index.js';
+
+// Import types that are used in the implementation
+import type {
+  PaginationParams,
+  SonarQubeProjectsResult,
+  SonarQubeIssuesResult,
+  IssuesParams,
+  SonarQubeMetricsResult,
+  ComponentMeasuresParams,
+  ComponentsMeasuresParams,
+  MeasuresHistoryParams,
+  SonarQubeComponentMeasuresResult,
+  SonarQubeComponentsMeasuresResult,
+  SonarQubeMeasuresHistoryResult,
+  SonarQubeHealthStatus,
+  SonarQubeSystemStatus,
+  SonarQubeQualityGatesResult,
+  SonarQubeQualityGate,
+  SonarQubeQualityGateStatus,
+  ProjectQualityGateParams,
+  SourceCodeParams,
+  ScmBlameParams,
+  SonarQubeSourceResult,
+  SonarQubeScmBlameResult,
+  HotspotSearchParams,
+  SonarQubeHotspot,
+  SonarQubeHotspotSearchResult,
+  SonarQubeHotspotDetails,
+  HotspotStatusUpdateParams,
+  ISonarQubeClient,
+} from './types/index.js';
+
+// Re-export all types for backward compatibility
+export type {
+  PaginationParams,
+  SeverityLevel,
+  SonarQubeProject,
+  SonarQubeProjectsResult,
+  SonarQubeIssue,
+  SonarQubeIssueComment,
+  SonarQubeIssueFlow,
+  SonarQubeIssueImpact,
+  SonarQubeIssueLocation,
+  SonarQubeMessageFormatting,
+  SonarQubeTextRange,
+  SonarQubeComponent,
+  SonarQubeRule,
+  SonarQubeUser,
+  SonarQubeFacet,
+  SonarQubeFacetValue,
+  SonarQubeIssuesResult,
+  IssuesParams,
+  SonarQubeMetric,
+  SonarQubeMetricsResult,
+  ComponentMeasuresParams,
+  ComponentsMeasuresParams,
+  MeasuresHistoryParams,
+  SonarQubeMeasure,
+  SonarQubeMeasureComponent,
+  SonarQubeComponentMeasuresResult,
+  SonarQubeComponentsMeasuresResult,
+  SonarQubeMeasuresHistoryResult,
+  SonarQubeHealthStatus,
+  SonarQubeSystemStatus,
+  SonarQubeQualityGateCondition,
+  SonarQubeQualityGate,
+  SonarQubeQualityGatesResult,
+  SonarQubeQualityGateStatus,
+  ProjectQualityGateParams,
+  SourceCodeParams,
+  ScmBlameParams,
+  SonarQubeLineIssue,
+  SonarQubeScmAuthor,
+  SonarQubeSourceLine,
+  SonarQubeSourceResult,
+  SonarQubeScmBlameResult,
+  HotspotSearchParams,
+  SonarQubeHotspot,
+  SonarQubeHotspotSearchResult,
+  SonarQubeHotspotDetails,
+  HotspotStatusUpdateParams,
+  ISonarQubeClient,
+} from './types/index.js';
 
 const logger = createLogger('sonarqube');
-
-/**
- * Type alias for severity levels
- */
-export type SeverityLevel = 'HIGH' | 'MEDIUM' | 'LOW';
 
 /**
  * Default SonarQube URL
@@ -14,745 +102,21 @@ export type SeverityLevel = 'HIGH' | 'MEDIUM' | 'LOW';
 const DEFAULT_SONARQUBE_URL = 'https://sonarcloud.io';
 
 /**
- * Interface for pagination parameters
- */
-export interface PaginationParams {
-  page?: number;
-  pageSize?: number;
-}
-
-/**
- * Interface for SonarQube project
- */
-export interface SonarQubeProject {
-  key: string;
-  name: string;
-  qualifier: string;
-  visibility: string;
-  lastAnalysisDate?: string;
-  revision?: string;
-  managed?: boolean;
-}
-
-/**
- * Interface for SonarQube issue impact
- */
-export interface SonarQubeIssueImpact {
-  softwareQuality: string;
-  severity: string;
-}
-
-/**
- * Interface for text range in SonarQube
- */
-export interface SonarQubeTextRange {
-  startLine: number;
-  endLine: number;
-  startOffset: number;
-  endOffset: number;
-}
-
-/**
- * Interface for message formatting in SonarQube
- */
-export interface SonarQubeMessageFormatting {
-  start: number;
-  end: number;
-  type: string;
-}
-
-/**
- * Interface for issue location in SonarQube
- */
-export interface SonarQubeIssueLocation {
-  textRange: SonarQubeTextRange;
-  msg: string;
-  msgFormattings?: SonarQubeMessageFormatting[];
-}
-
-/**
- * Interface for issue flow in SonarQube
- */
-export interface SonarQubeIssueFlow {
-  locations: SonarQubeIssueLocation[];
-}
-
-/**
- * Interface for issue comment in SonarQube
- */
-export interface SonarQubeIssueComment {
-  key: string;
-  login: string;
-  htmlText: string;
-  markdown: string;
-  updatable: boolean;
-  createdAt: string;
-}
-
-/**
- * Interface for SonarQube issue
- */
-export interface SonarQubeIssue {
-  key: string;
-  rule: string;
-  component: string;
-  project: string;
-  line?: number;
-  hash?: string;
-  textRange?: SonarQubeTextRange;
-  message: string;
-  messageFormattings?: SonarQubeMessageFormatting[];
-  status: string;
-  issueStatus?: string;
-  effort?: string;
-  debt?: string;
-  author?: string;
-  severity?: string;
-  tags: string[];
-  creationDate: string;
-  updateDate: string;
-  type?: string;
-  cleanCodeAttribute?: string;
-  cleanCodeAttributeCategory?: string;
-  prioritizedRule?: boolean;
-  impacts?: SonarQubeIssueImpact[];
-  comments?: SonarQubeIssueComment[];
-  transitions?: string[];
-  actions?: string[];
-  flows?: SonarQubeIssueFlow[];
-  quickFixAvailable?: boolean;
-  ruleDescriptionContextKey?: string;
-  codeVariants?: string[];
-}
-
-/**
- * Interface for SonarQube component
- */
-export interface SonarQubeComponent {
-  key: string;
-  enabled?: boolean;
-  qualifier: string;
-  name: string;
-  longName?: string;
-  path?: string;
-}
-
-/**
- * Interface for SonarQube rule
- */
-export interface SonarQubeRule {
-  key: string;
-  name: string;
-  status: string;
-  lang: string;
-  langName: string;
-}
-
-/**
- * Interface for SonarQube user
- */
-export interface SonarQubeUser {
-  login: string;
-  name: string;
-  active: boolean;
-  avatar?: string;
-}
-
-/**
- * Interface for SonarQube facet value
- */
-export interface SonarQubeFacetValue {
-  val: string;
-  count: number;
-}
-
-/**
- * Interface for SonarQube facet
- */
-export interface SonarQubeFacet {
-  property: string;
-  values: SonarQubeFacetValue[];
-}
-
-/**
- * Interface for SonarQube issues result
- */
-export interface SonarQubeIssuesResult {
-  issues: SonarQubeIssue[];
-  components: SonarQubeComponent[];
-  rules: SonarQubeRule[];
-  users?: SonarQubeUser[];
-  facets?: SonarQubeFacet[];
-  paging: {
-    pageIndex: number;
-    pageSize: number;
-    total: number;
-  };
-}
-
-/**
- * Interface for SonarQube projects result - Clean abstraction for consumers
- */
-export interface SonarQubeProjectsResult {
-  projects: SonarQubeProject[];
-  paging: {
-    pageIndex: number;
-    pageSize: number;
-    total: number;
-  };
-}
-
-/**
- * Interface for get issues parameters
- */
-export interface IssuesParams extends PaginationParams {
-  // Component filters
-  projectKey?: string;
-  componentKeys?: string[];
-  components?: string[];
-  projects?: string[];
-  onComponentOnly?: boolean;
-
-  // Branch and PR
-  branch?: string;
-  pullRequest?: string;
-
-  // Issue filters
-  issues?: string[];
-  severities?: ('INFO' | 'MINOR' | 'MAJOR' | 'CRITICAL' | 'BLOCKER')[];
-  statuses?: ('OPEN' | 'CONFIRMED' | 'REOPENED' | 'RESOLVED' | 'CLOSED')[];
-  resolutions?: ('FALSE-POSITIVE' | 'WONTFIX' | 'FIXED' | 'REMOVED')[];
-  resolved?: boolean;
-  types?: ('CODE_SMELL' | 'BUG' | 'VULNERABILITY' | 'SECURITY_HOTSPOT')[];
-
-  // Clean Code taxonomy
-  cleanCodeAttributeCategories?: ('ADAPTABLE' | 'CONSISTENT' | 'INTENTIONAL' | 'RESPONSIBLE')[];
-  impactSeverities?: SeverityLevel[];
-  impactSoftwareQualities?: ('MAINTAINABILITY' | 'RELIABILITY' | 'SECURITY')[];
-  issueStatuses?: ('OPEN' | 'CONFIRMED' | 'RESOLVED' | 'REOPENED' | 'CLOSED')[];
-
-  // Rules and tags
-  rules?: string[];
-  tags?: string[];
-
-  // Date filters
-  createdAfter?: string;
-  createdBefore?: string;
-  createdAt?: string;
-  createdInLast?: string;
-
-  // Assignment
-  assigned?: boolean;
-  assignees?: string[];
-  author?: string;
-  authors?: string[];
-
-  // Security standards
-  cwe?: string[];
-  owaspTop10?: string[];
-  owaspTop10v2021?: string[];
-  sansTop25?: string[];
-  sonarsourceSecurity?: string[];
-  sonarsourceSecurityCategory?: string[];
-
-  // Languages
-  languages?: string[];
-
-  // Facets
-  facets?: string[];
-  facetMode?: 'effort' | 'count';
-
-  // New code
-  sinceLeakPeriod?: boolean;
-  inNewCodePeriod?: boolean;
-
-  // Sorting
-  s?: string;
-  asc?: boolean;
-
-  // Additional fields
-  additionalFields?: string[];
-
-  // Deprecated
-  hotspots?: boolean;
-  severity?: 'INFO' | 'MINOR' | 'MAJOR' | 'CRITICAL' | 'BLOCKER';
-}
-
-/**
- * Interface for component measures parameters
- */
-export interface ComponentMeasuresParams {
-  component: string;
-  metricKeys: string[];
-  additionalFields?: string[];
-  branch?: string;
-  pullRequest?: string;
-  period?: string;
-}
-
-/**
- * Interface for components measures parameters
- */
-export interface ComponentsMeasuresParams extends PaginationParams {
-  componentKeys: string[] | string;
-  metricKeys: string[] | string;
-  additionalFields?: string[];
-  branch?: string;
-  pullRequest?: string;
-  period?: string;
-}
-
-/**
- * Interface for measures history parameters
- */
-export interface MeasuresHistoryParams extends PaginationParams {
-  component: string;
-  metrics: string[];
-  from?: string;
-  to?: string;
-  branch?: string;
-  pullRequest?: string;
-}
-
-/**
- * Interface for SonarQube metric
- */
-export interface SonarQubeMetric {
-  id: string;
-  key: string;
-  name: string;
-  description: string;
-  domain: string;
-  type: string;
-  direction: number;
-  qualitative: boolean;
-  hidden: boolean;
-  custom: boolean;
-}
-
-/**
- * Interface for SonarQube metrics result
- */
-export interface SonarQubeMetricsResult {
-  metrics: SonarQubeMetric[];
-  paging: {
-    pageIndex: number;
-    pageSize: number;
-    total: number;
-  };
-}
-
-/**
- * Interface for SonarQube measure
- */
-export interface SonarQubeMeasure {
-  metric: string;
-  value?: string;
-  period?: {
-    index: number;
-    value: string;
-  };
-  bestValue?: boolean;
-}
-
-/**
- * Interface for SonarQube measure component
- */
-export interface SonarQubeMeasureComponent {
-  key: string;
-  name: string;
-  qualifier: string;
-  measures: SonarQubeMeasure[];
-  periods?: Array<{
-    index: number;
-    mode: string;
-    date: string;
-    parameter?: string;
-  }>;
-}
-
-/**
- * Interface for SonarQube component with measures result
- */
-export interface SonarQubeComponentMeasuresResult {
-  component: SonarQubeMeasureComponent;
-  metrics: SonarQubeMetric[];
-  period?: {
-    index: number;
-    mode: string;
-    date: string;
-    parameter?: string;
-  };
-}
-
-/**
- * Interface for SonarQube components with measures result
- */
-export interface SonarQubeComponentsMeasuresResult {
-  components: SonarQubeMeasureComponent[];
-  metrics: SonarQubeMetric[];
-  paging: {
-    pageIndex: number;
-    pageSize: number;
-    total: number;
-  };
-  period?: {
-    index: number;
-    mode: string;
-    date: string;
-    parameter?: string;
-  };
-}
-
-/**
- * Interface for SonarQube measures history result
- */
-export interface SonarQubeMeasuresHistoryResult {
-  paging: {
-    pageIndex: number;
-    pageSize: number;
-    total: number;
-  };
-  measures: {
-    metric: string;
-    history: {
-      date: string;
-      value?: string;
-    }[];
-  }[];
-}
-
-/**
- * Interface for SonarQube health status
- */
-export interface SonarQubeHealthStatus {
-  health: 'GREEN' | 'YELLOW' | 'RED';
-  causes: string[];
-}
-
-/**
- * Interface for SonarQube system status
- */
-export interface SonarQubeSystemStatus {
-  id: string;
-  version: string;
-  status:
-    | 'UP'
-    | 'DOWN'
-    | 'STARTING'
-    | 'RESTARTING'
-    | 'DB_MIGRATION_NEEDED'
-    | 'DB_MIGRATION_RUNNING';
-}
-
-/**
- * Interface for SonarQube quality gate condition
- */
-export interface SonarQubeQualityGateCondition {
-  id: string;
-  metric: string;
-  op: string;
-  error: string;
-}
-
-/**
- * Interface for SonarQube quality gate
- */
-export interface SonarQubeQualityGate {
-  id: string;
-  name: string;
-  isDefault?: boolean;
-  isBuiltIn?: boolean;
-  conditions?: SonarQubeQualityGateCondition[];
-}
-
-/**
- * Interface for SonarQube quality gates list result
- */
-export interface SonarQubeQualityGatesResult {
-  qualitygates: SonarQubeQualityGate[];
-  default: string;
-  actions?: {
-    create?: boolean;
-  };
-}
-
-/**
- * Interface for SonarQube quality gate status
- */
-export interface SonarQubeQualityGateStatus {
-  projectStatus: {
-    status: 'OK' | 'WARN' | 'ERROR' | 'NONE';
-    conditions: Array<{
-      status: 'OK' | 'WARN' | 'ERROR';
-      metricKey: string;
-      comparator: string;
-      errorThreshold: string;
-      actualValue: string;
-    }>;
-    periods?: Array<{
-      index: number;
-      mode: string;
-      date: string;
-      parameter?: string;
-    }>;
-    ignoredConditions: boolean;
-  };
-}
-
-/**
- * Interface for project quality gate params
- */
-export interface ProjectQualityGateParams {
-  projectKey: string;
-  branch?: string;
-  pullRequest?: string;
-}
-
-/**
- * Interface for source code location parameters
- */
-export interface SourceCodeParams {
-  key: string;
-  from?: number;
-  to?: number;
-  branch?: string;
-  pullRequest?: string;
-}
-
-/**
- * Interface for SCM blame parameters
- */
-export interface ScmBlameParams {
-  key: string;
-  from?: number;
-  to?: number;
-  branch?: string;
-  pullRequest?: string;
-}
-
-/**
- * Interface for line issue in source code
- */
-export interface SonarQubeLineIssue {
-  line: number;
-  issues: SonarQubeIssue[];
-}
-
-/**
- * Interface for SCM author information
- */
-export interface SonarQubeScmAuthor {
-  revision: string;
-  date: string;
-  author: string;
-}
-
-/**
- * Interface for source code line with annotations
- */
-export interface SonarQubeSourceLine {
-  line: number;
-  code: string;
-  scmAuthor?: string;
-  scmDate?: string;
-  scmRevision?: string;
-  duplicated?: boolean;
-  isNew?: boolean;
-  lineHits?: number;
-  conditions?: number;
-  coveredConditions?: number;
-  highlightedText?: string;
-  issues?: SonarQubeIssue[];
-}
-
-/**
- * Interface for source code result
- */
-export interface SonarQubeSourceResult {
-  component: {
-    key: string;
-    path?: string;
-    qualifier: string;
-    name: string;
-    longName?: string;
-    language?: string;
-  };
-  sources: SonarQubeSourceLine[];
-}
-
-/**
- * Interface for SCM blame result
- */
-export interface SonarQubeScmBlameResult {
-  component: {
-    key: string;
-    path?: string;
-    qualifier: string;
-    name: string;
-    longName?: string;
-    language?: string;
-  };
-  sources: {
-    [lineNumber: string]: SonarQubeScmAuthor;
-  };
-}
-
-/**
- * Interface for hotspot search parameters
- */
-export interface HotspotSearchParams extends PaginationParams {
-  projectKey?: string;
-  branch?: string;
-  pullRequest?: string;
-  status?: 'TO_REVIEW' | 'REVIEWED';
-  resolution?: 'FIXED' | 'SAFE';
-  files?: string[];
-  assignedToMe?: boolean;
-  sinceLeakPeriod?: boolean;
-  inNewCodePeriod?: boolean;
-}
-
-/**
- * Interface for security hotspot
- */
-export interface SonarQubeHotspot {
-  key: string;
-  component: string;
-  project: string;
-  securityCategory: string;
-  vulnerabilityProbability: SeverityLevel;
-  status: 'TO_REVIEW' | 'REVIEWED';
-  resolution?: 'FIXED' | 'SAFE';
-  line: number;
-  message: string;
-  assignee?: string;
-  author?: string;
-  creationDate: string;
-  updateDate: string;
-  textRange?: {
-    startLine: number;
-    endLine: number;
-    startOffset: number;
-    endOffset: number;
-  };
-  flows?: Array<{
-    locations: Array<{
-      component: string;
-      textRange: {
-        startLine: number;
-        endLine: number;
-        startOffset: number;
-        endOffset: number;
-      };
-      msg: string;
-    }>;
-  }>;
-  ruleKey?: string;
-}
-
-/**
- * Interface for hotspot search result
- */
-export interface SonarQubeHotspotSearchResult {
-  hotspots: SonarQubeHotspot[];
-  components?: Array<{
-    key: string;
-    qualifier: string;
-    name: string;
-    longName?: string;
-    path?: string;
-  }>;
-  paging: {
-    pageIndex: number;
-    pageSize: number;
-    total: number;
-  };
-}
-
-/**
- * Interface for hotspot details
- */
-export interface SonarQubeHotspotDetails extends SonarQubeHotspot {
-  rule: {
-    key: string;
-    name: string;
-    securityCategory: string;
-    vulnerabilityProbability: SeverityLevel;
-  };
-  changelog?: Array<{
-    user: string;
-    userName?: string;
-    creationDate: string;
-    diffs: Array<{
-      key: string;
-      oldValue?: string;
-      newValue?: string;
-    }>;
-  }>;
-  comment?: Array<{
-    key: string;
-    login: string;
-    htmlText: string;
-    markdown?: string;
-    createdAt: string;
-  }>;
-  users?: Array<{
-    login: string;
-    name: string;
-    active: boolean;
-  }>;
-}
-
-/**
- * Interface for hotspot status update parameters
- */
-export interface HotspotStatusUpdateParams {
-  hotspot: string;
-  status: 'TO_REVIEW' | 'REVIEWED';
-  resolution?: 'FIXED' | 'SAFE';
-  comment?: string;
-}
-
-/**
- * Interface for SonarQube client
- */
-export interface ISonarQubeClient {
-  listProjects(params?: PaginationParams): Promise<SonarQubeProjectsResult>;
-  getIssues(params: IssuesParams): Promise<SonarQubeIssuesResult>;
-  getMetrics(params?: PaginationParams): Promise<SonarQubeMetricsResult>;
-  getHealth(): Promise<SonarQubeHealthStatus>;
-  getStatus(): Promise<SonarQubeSystemStatus>;
-  ping(): Promise<string>;
-  getComponentMeasures(params: ComponentMeasuresParams): Promise<SonarQubeComponentMeasuresResult>;
-  getComponentsMeasures(
-    params: ComponentsMeasuresParams
-  ): Promise<SonarQubeComponentsMeasuresResult>;
-  getMeasuresHistory(params: MeasuresHistoryParams): Promise<SonarQubeMeasuresHistoryResult>;
-
-  // Quality Gates API methods
-  listQualityGates(): Promise<SonarQubeQualityGatesResult>;
-  getQualityGate(id: string): Promise<SonarQubeQualityGate>;
-  getProjectQualityGateStatus(
-    params: ProjectQualityGateParams
-  ): Promise<SonarQubeQualityGateStatus>;
-
-  // Source Code API methods
-  getSourceCode(params: SourceCodeParams): Promise<SonarQubeSourceResult>;
-  getScmBlame(params: ScmBlameParams): Promise<SonarQubeScmBlameResult>;
-
-  // Security Hotspots API methods
-  hotspots(params: HotspotSearchParams): Promise<SonarQubeHotspotSearchResult>;
-  hotspot(hotspotKey: string): Promise<SonarQubeHotspotDetails>;
-  updateHotspotStatus(params: HotspotStatusUpdateParams): Promise<void>;
-}
-
-/**
  * SonarQube client for interacting with the SonarQube API
  */
 export class SonarQubeClient implements ISonarQubeClient {
   private readonly webApiClient: WebApiClient;
   private readonly organization: string | null;
+
+  // Domain modules
+  private readonly projectsDomain: ProjectsDomain;
+  private readonly issuesDomain: IssuesDomain;
+  private readonly metricsDomain: MetricsDomain;
+  private readonly measuresDomain: MeasuresDomain;
+  private readonly systemDomain: SystemDomain;
+  private readonly qualityGatesDomain: QualityGatesDomain;
+  private readonly sourceCodeDomain: SourceCodeDomain;
+  private readonly hotspotsDomain: HotspotsDomain;
 
   /**
    * Creates a new SonarQube client
@@ -767,6 +131,20 @@ export class SonarQubeClient implements ISonarQubeClient {
       organization ? { organization } : undefined
     );
     this.organization = organization ?? null;
+
+    // Initialize domain modules
+    this.projectsDomain = new ProjectsDomain(this.webApiClient, this.organization);
+    this.issuesDomain = new IssuesDomain(this.webApiClient, this.organization);
+    this.metricsDomain = new MetricsDomain(this.webApiClient, this.organization);
+    this.measuresDomain = new MeasuresDomain(this.webApiClient, this.organization);
+    this.systemDomain = new SystemDomain(this.webApiClient, this.organization);
+    this.qualityGatesDomain = new QualityGatesDomain(this.webApiClient, this.organization);
+    this.sourceCodeDomain = new SourceCodeDomain(
+      this.webApiClient,
+      this.organization,
+      this.issuesDomain
+    );
+    this.hotspotsDomain = new HotspotsDomain(this.webApiClient, this.organization);
   }
 
   /**
@@ -791,6 +169,21 @@ export class SonarQubeClient implements ISonarQubeClient {
       organization ? { organization } : undefined
     );
     client.organization = organization ?? null;
+
+    // Initialize domain modules
+    client.projectsDomain = new ProjectsDomain(client.webApiClient, client.organization);
+    client.issuesDomain = new IssuesDomain(client.webApiClient, client.organization);
+    client.metricsDomain = new MetricsDomain(client.webApiClient, client.organization);
+    client.measuresDomain = new MeasuresDomain(client.webApiClient, client.organization);
+    client.systemDomain = new SystemDomain(client.webApiClient, client.organization);
+    client.qualityGatesDomain = new QualityGatesDomain(client.webApiClient, client.organization);
+    client.sourceCodeDomain = new SourceCodeDomain(
+      client.webApiClient,
+      client.organization,
+      client.issuesDomain
+    );
+    client.hotspotsDomain = new HotspotsDomain(client.webApiClient, client.organization);
+
     return client;
   }
 
@@ -813,6 +206,21 @@ export class SonarQubeClient implements ISonarQubeClient {
       organization ? { organization } : undefined
     );
     client.organization = organization ?? null;
+
+    // Initialize domain modules
+    client.projectsDomain = new ProjectsDomain(client.webApiClient, client.organization);
+    client.issuesDomain = new IssuesDomain(client.webApiClient, client.organization);
+    client.metricsDomain = new MetricsDomain(client.webApiClient, client.organization);
+    client.measuresDomain = new MeasuresDomain(client.webApiClient, client.organization);
+    client.systemDomain = new SystemDomain(client.webApiClient, client.organization);
+    client.qualityGatesDomain = new QualityGatesDomain(client.webApiClient, client.organization);
+    client.sourceCodeDomain = new SourceCodeDomain(
+      client.webApiClient,
+      client.organization,
+      client.issuesDomain
+    );
+    client.hotspotsDomain = new HotspotsDomain(client.webApiClient, client.organization);
+
     return client;
   }
 
@@ -822,39 +230,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the list of projects
    */
   async listProjects(params: PaginationParams = {}): Promise<SonarQubeProjectsResult> {
-    const { page, pageSize } = params;
-    logger.debug('Listing projects', { page, pageSize, organization: this.organization });
-
-    try {
-      const builder = this.webApiClient.projects.search();
-
-      if (page !== undefined) {
-        builder.page(page);
-      }
-      if (pageSize !== undefined) {
-        builder.pageSize(pageSize);
-      }
-
-      const response = await builder.execute();
-      logger.debug('Projects retrieved successfully', { count: response.components.length });
-
-      // Transform to our interface
-      return {
-        projects: response.components.map((component) => ({
-          key: component.key,
-          name: component.name,
-          qualifier: component.qualifier,
-          visibility: component.visibility,
-          lastAnalysisDate: component.lastAnalysisDate,
-          revision: component.revision,
-          managed: component.managed,
-        })),
-        paging: response.paging,
-      };
-    } catch (error) {
-      logger.error('Failed to list projects', error);
-      throw error;
-    }
+    return this.projectsDomain.listProjects(params);
   }
 
   /**
@@ -863,34 +239,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the list of issues
    */
   async getIssues(params: IssuesParams): Promise<SonarQubeIssuesResult> {
-    const { page, pageSize } = params;
-    const builder = this.webApiClient.issues.search();
-
-    // Apply all filters using helper methods
-    this.applyComponentFilters(builder, params);
-    this.applyIssueFilters(builder, params);
-    this.applyDateAndAssignmentFilters(builder, params);
-    this.applySecurityAndMetadataFilters(builder, params);
-
-    // Add pagination
-    if (page !== undefined) {
-      builder.page(page);
-    }
-    if (pageSize !== undefined) {
-      builder.pageSize(pageSize);
-    }
-
-    const response = await builder.execute();
-
-    // Transform to our interface
-    return {
-      issues: response.issues as SonarQubeIssue[],
-      components: response.components ?? [],
-      rules: (response.rules ?? []) as SonarQubeRule[],
-      users: response.users,
-      facets: response.facets,
-      paging: response.paging ?? { pageIndex: 1, pageSize: 100, total: 0 },
-    };
+    return this.issuesDomain.getIssues(params);
   }
 
   /**
@@ -899,28 +248,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the list of metrics
    */
   async getMetrics(params: PaginationParams = {}): Promise<SonarQubeMetricsResult> {
-    const { page, pageSize } = params;
-
-    const response = await this.webApiClient.metrics.search({
-      p: page,
-      ps: pageSize,
-    });
-
-    // The API might return paging info
-    const paging = (
-      response as unknown as {
-        paging?: { pageIndex: number; pageSize: number; total: number };
-      }
-    ).paging;
-
-    return {
-      metrics: response.metrics as SonarQubeMetric[],
-      paging: paging ?? {
-        pageIndex: page ?? 1,
-        pageSize: pageSize ?? 100,
-        total: response.metrics.length,
-      },
-    };
+    return this.metricsDomain.getMetrics(params);
   }
 
   /**
@@ -928,11 +256,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the health status
    */
   async getHealth(): Promise<SonarQubeHealthStatus> {
-    const response = await this.webApiClient.system.health();
-    return {
-      health: response.health,
-      causes: response.causes ?? [],
-    };
+    return this.systemDomain.getHealth();
   }
 
   /**
@@ -940,12 +264,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the system status
    */
   async getStatus(): Promise<SonarQubeSystemStatus> {
-    const response = await this.webApiClient.system.status();
-    return {
-      id: response.id,
-      version: response.version,
-      status: response.status,
-    };
+    return this.systemDomain.getStatus();
   }
 
   /**
@@ -953,7 +272,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the ping response
    */
   async ping(): Promise<string> {
-    return this.webApiClient.system.ping();
+    return this.systemDomain.ping();
   }
 
   /**
@@ -964,17 +283,7 @@ export class SonarQubeClient implements ISonarQubeClient {
   async getComponentMeasures(
     params: ComponentMeasuresParams
   ): Promise<SonarQubeComponentMeasuresResult> {
-    const { component, metricKeys, additionalFields, branch, pullRequest } = params;
-
-    const response = await this.webApiClient.measures.component({
-      component,
-      metricKeys: Array.isArray(metricKeys) ? metricKeys : [metricKeys as string],
-      additionalFields: additionalFields as MeasuresAdditionalField[] | undefined,
-      branch,
-      pullRequest,
-    });
-
-    return response as SonarQubeComponentMeasuresResult;
+    return this.measuresDomain.getComponentMeasures(params);
   }
 
   /**
@@ -992,78 +301,7 @@ export class SonarQubeClient implements ISonarQubeClient {
   async getComponentsMeasures(
     params: ComponentsMeasuresParams
   ): Promise<SonarQubeComponentsMeasuresResult> {
-    const { componentKeys, metricKeys, additionalFields, branch, pullRequest, page, pageSize } =
-      params;
-
-    // Use Promise.all to fetch measures for multiple components
-    // Note: This results in N+1 API calls (one per component)
-    let componentKeysArray: string[];
-    if (Array.isArray(componentKeys)) {
-      componentKeysArray = componentKeys;
-    } else if (typeof componentKeys === 'string' && componentKeys.includes(',')) {
-      componentKeysArray = componentKeys.split(',');
-    } else {
-      componentKeysArray = [componentKeys];
-    }
-
-    let metricKeysArray: string[];
-    if (Array.isArray(metricKeys)) {
-      metricKeysArray = metricKeys;
-    } else if (typeof metricKeys === 'string' && metricKeys.includes(',')) {
-      metricKeysArray = metricKeys.split(',');
-    } else {
-      metricKeysArray = [metricKeys];
-    }
-
-    const componentsPromises = componentKeysArray.map(async (componentKey: string) => {
-      const result = await this.webApiClient.measures.component({
-        component: componentKey,
-        metricKeys: metricKeysArray,
-        additionalFields: additionalFields as MeasuresAdditionalField[] | undefined,
-        branch,
-        pullRequest,
-      });
-
-      const component = result.component ?? result;
-      return {
-        key: component.key,
-        name: component.name,
-        qualifier: component.qualifier,
-        measures: component.measures ?? [],
-        periods: (component as unknown as SonarQubeMeasureComponent).periods,
-      };
-    });
-
-    const components = await Promise.all(componentsPromises);
-
-    // Get metrics from the first component response (they should be the same for all)
-    const firstResult = await this.webApiClient.measures.component({
-      component: componentKeysArray[0],
-      metricKeys: metricKeysArray,
-      additionalFields: additionalFields as MeasuresAdditionalField[] | undefined,
-      branch,
-      pullRequest,
-    });
-
-    // Apply pagination manually
-    const startIndex = ((page ?? 1) - 1) * (pageSize ?? 100);
-    const endIndex = startIndex + (pageSize ?? 100);
-    const paginatedComponents = components.slice(startIndex, endIndex);
-
-    return {
-      components: paginatedComponents,
-      metrics: (firstResult.metrics ?? []) as SonarQubeMetric[],
-      paging: {
-        pageIndex: page ?? 1,
-        pageSize: pageSize ?? 100,
-        total: components.length,
-      },
-      period: (
-        firstResult as unknown as {
-          period?: { index: number; mode: string; date: string; parameter?: string };
-        }
-      ).period,
-    };
+    return this.measuresDomain.getComponentsMeasures(params);
   }
 
   /**
@@ -1072,37 +310,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the measures history result
    */
   async getMeasuresHistory(params: MeasuresHistoryParams): Promise<SonarQubeMeasuresHistoryResult> {
-    const { component, metrics, from, to, branch, pullRequest, page, pageSize } = params;
-
-    const builder = this.webApiClient.measures.searchHistory(
-      component,
-      Array.isArray(metrics) ? metrics : [metrics as string]
-    );
-
-    if (from) {
-      builder.from(from);
-    }
-    if (to) {
-      builder.to(to);
-    }
-    if (branch) {
-      builder.withBranch(branch);
-    }
-    if (pullRequest) {
-      builder.withPullRequest(pullRequest);
-    }
-    if (page !== undefined) {
-      builder.page(page);
-    }
-    if (pageSize !== undefined) {
-      builder.pageSize(pageSize);
-    }
-
-    const response = await builder.execute();
-    return {
-      ...response,
-      paging: response.paging ?? { pageIndex: 1, pageSize: 100, total: 0 },
-    };
+    return this.measuresDomain.getMeasuresHistory(params);
   }
 
   /**
@@ -1110,12 +318,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the list of quality gates
    */
   async listQualityGates(): Promise<SonarQubeQualityGatesResult> {
-    const response = await this.webApiClient.qualityGates.list();
-    return {
-      qualitygates: response.qualitygates as SonarQubeQualityGate[],
-      default: response.default ?? '',
-      actions: (response as unknown as { actions?: { create?: boolean } }).actions,
-    };
+    return this.qualityGatesDomain.listQualityGates();
   }
 
   /**
@@ -1124,14 +327,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise with the quality gate details
    */
   async getQualityGate(id: string): Promise<SonarQubeQualityGate> {
-    const response = await this.webApiClient.qualityGates.get({ id });
-    return {
-      id: response.id,
-      name: response.name,
-      isDefault: response.isDefault,
-      isBuiltIn: response.isBuiltIn,
-      conditions: response.conditions as unknown as SonarQubeQualityGateCondition[],
-    };
+    return this.qualityGatesDomain.getQualityGate(id);
   }
 
   /**
@@ -1142,15 +338,7 @@ export class SonarQubeClient implements ISonarQubeClient {
   async getProjectQualityGateStatus(
     params: ProjectQualityGateParams
   ): Promise<SonarQubeQualityGateStatus> {
-    const { projectKey, branch, pullRequest } = params;
-
-    const response = await this.webApiClient.qualityGates.getProjectStatus({
-      projectKey,
-      branch,
-      pullRequest,
-    });
-
-    return response as unknown as SonarQubeQualityGateStatus;
+    return this.qualityGatesDomain.getProjectQualityGateStatus(params);
   }
 
   /**
@@ -1313,289 +501,7 @@ export class SonarQubeClient implements ISonarQubeClient {
    * @returns Promise that resolves when the update is complete
    */
   async updateHotspotStatus(params: HotspotStatusUpdateParams): Promise<void> {
-    const { hotspot, status, resolution, comment } = params;
-
-    await this.webApiClient.hotspots.changeStatus({
-      hotspot,
-      status,
-      ...(resolution && { resolution }),
-      ...(comment && { comment }),
-    });
-  }
-
-  /**
-   * Apply component filters to the search builder
-   */
-  private applyComponentFilters(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    builder: any,
-    params: Pick<
-      IssuesParams,
-      | 'projectKey'
-      | 'componentKeys'
-      | 'components'
-      | 'projects'
-      | 'onComponentOnly'
-      | 'branch'
-      | 'pullRequest'
-    >
-  ): void {
-    const {
-      projectKey,
-      componentKeys,
-      components,
-      projects,
-      onComponentOnly,
-      branch,
-      pullRequest,
-    } = params;
-
-    if (projectKey) {
-      builder.withProjects([projectKey]);
-    }
-    if (componentKeys) {
-      builder.withComponents(componentKeys);
-    }
-    if (components) {
-      builder.withComponents(components);
-    }
-    if (projects) {
-      builder.withProjects(projects);
-    }
-    if (onComponentOnly) {
-      builder.onComponentOnly();
-    }
-    if (branch) {
-      builder.onBranch(branch);
-    }
-    if (pullRequest) {
-      builder.onPullRequest(pullRequest);
-    }
-  }
-
-  /**
-   * Apply issue filters to the search builder
-   */
-  private applyIssueFilters(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    builder: any,
-    params: Pick<
-      IssuesParams,
-      | 'issues'
-      | 'severities'
-      | 'severity'
-      | 'statuses'
-      | 'resolutions'
-      | 'resolved'
-      | 'types'
-      | 'cleanCodeAttributeCategories'
-      | 'impactSeverities'
-      | 'impactSoftwareQualities'
-      | 'issueStatuses'
-    >
-  ): void {
-    const {
-      issues,
-      severities,
-      severity,
-      statuses,
-      resolutions,
-      resolved,
-      types,
-      cleanCodeAttributeCategories,
-      impactSeverities,
-      impactSoftwareQualities,
-      issueStatuses,
-    } = params;
-
-    if (issues) {
-      builder.withIssues(issues);
-    }
-    if (severities) {
-      builder.withSeverities(severities);
-    }
-    if (severity) {
-      builder.withSeverities([severity]);
-    }
-    if (statuses) {
-      builder.withStatuses(statuses);
-    }
-    if (resolutions) {
-      builder.withResolutions(resolutions);
-    }
-    if (resolved !== undefined) {
-      if (resolved) {
-        builder.onlyResolved();
-      } else {
-        builder.onlyUnresolved();
-      }
-    }
-    if (types) {
-      builder.withTypes(types);
-    }
-    if (cleanCodeAttributeCategories) {
-      builder.withCleanCodeAttributeCategories(cleanCodeAttributeCategories);
-    }
-    if (impactSeverities) {
-      builder.withImpactSeverities(impactSeverities);
-    }
-    if (impactSoftwareQualities) {
-      builder.withImpactSoftwareQualities(impactSoftwareQualities);
-    }
-    if (issueStatuses) {
-      builder.withIssueStatuses(issueStatuses);
-    }
-  }
-
-  /**
-   * Apply date and assignment filters to the search builder
-   */
-  private applyDateAndAssignmentFilters(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    builder: any,
-    params: Pick<
-      IssuesParams,
-      | 'createdAfter'
-      | 'createdBefore'
-      | 'createdAt'
-      | 'createdInLast'
-      | 'assigned'
-      | 'assignees'
-      | 'author'
-      | 'authors'
-    >
-  ): void {
-    const {
-      createdAfter,
-      createdBefore,
-      createdAt,
-      createdInLast,
-      assigned,
-      assignees,
-      author,
-      authors,
-    } = params;
-
-    if (createdAfter) {
-      builder.createdAfter(createdAfter);
-    }
-    if (createdBefore) {
-      builder.createdBefore(createdBefore);
-    }
-    if (createdAt) {
-      builder.createdAt(createdAt);
-    }
-    if (createdInLast) {
-      builder.createdInLast(createdInLast);
-    }
-    if (assigned !== undefined) {
-      if (assigned) {
-        builder.onlyAssigned();
-      } else {
-        builder.onlyUnassigned();
-      }
-    }
-    if (assignees) {
-      builder.assignedToAny(assignees);
-    }
-    if (author) {
-      builder.byAuthor(author);
-    }
-    if (authors) {
-      builder.byAuthors(authors);
-    }
-  }
-
-  /**
-   * Apply security and metadata filters to the search builder
-   */
-  private applySecurityAndMetadataFilters(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    builder: any,
-    params: Pick<
-      IssuesParams,
-      | 'cwe'
-      | 'owaspTop10'
-      | 'owaspTop10v2021'
-      | 'sansTop25'
-      | 'sonarsourceSecurity'
-      | 'sonarsourceSecurityCategory'
-      | 'rules'
-      | 'tags'
-      | 'languages'
-      | 'facets'
-      | 'facetMode'
-      | 'sinceLeakPeriod'
-      | 'inNewCodePeriod'
-      | 's'
-      | 'asc'
-      | 'additionalFields'
-    >
-  ): void {
-    const {
-      cwe,
-      owaspTop10,
-      owaspTop10v2021,
-      sansTop25,
-      sonarsourceSecurity,
-      sonarsourceSecurityCategory,
-      rules,
-      tags,
-      languages,
-      facets,
-      facetMode,
-      sinceLeakPeriod,
-      inNewCodePeriod,
-      s,
-      asc,
-      additionalFields,
-    } = params;
-
-    if (cwe) {
-      builder.withCwe(cwe);
-    }
-    if (owaspTop10) {
-      builder.withOwaspTop10(owaspTop10);
-    }
-    if (owaspTop10v2021) {
-      builder.withOwaspTop10v2021(owaspTop10v2021);
-    }
-    if (sansTop25) {
-      builder.withSansTop25(sansTop25);
-    }
-    if (sonarsourceSecurity) {
-      builder.withSonarSourceSecurity(sonarsourceSecurity);
-    }
-    if (sonarsourceSecurityCategory) {
-      builder.withSonarSourceSecurityNew(sonarsourceSecurityCategory);
-    }
-    if (rules) {
-      builder.withRules(rules);
-    }
-    if (tags) {
-      builder.withTags(tags);
-    }
-    if (languages) {
-      builder.withLanguages(languages);
-    }
-    if (facets) {
-      builder.withFacets(facets);
-    }
-    if (facetMode) {
-      builder.withFacetMode(facetMode);
-    }
-    if (sinceLeakPeriod) {
-      builder.sinceLeakPeriod();
-    }
-    if (inNewCodePeriod) {
-      builder.inNewCodePeriod();
-    }
-    if (s) {
-      builder.sortBy(s, asc);
-    }
-    if (additionalFields) {
-      builder.withAdditionalFields(additionalFields);
-    }
+    return this.hotspotsDomain.updateHotspotStatus(params);
   }
 }
