+++ src/types/common.ts
@@ -1,0 +1,12 @@
+/**
+ * Type alias for severity levels
+ */
+export type SeverityLevel = 'HIGH' | 'MEDIUM' | 'LOW';
+
+/**
+ * Interface for pagination parameters
+ */
+export interface PaginationParams {
+  page?: number;
+  pageSize?: number;
+}
