+++ src/types/projects.ts
@@ -1,0 +1,24 @@
+/**
+ * Interface for SonarQube project
+ */
+export interface SonarQubeProject {
+  key: string;
+  name: string;
+  qualifier: string;
+  visibility: string;
+  lastAnalysisDate?: string;
+  revision?: string;
+  managed?: boolean;
+}
+
+/**
+ * Interface for SonarQube projects result - Clean abstraction for consumers
+ */
+export interface SonarQubeProjectsResult {
+  projects: SonarQubeProject[];
+  paging: {
+    pageIndex: number;
+    pageSize: number;
+    total: number;
+  };
+}
